# doc/api.yaml
openapi: 3.0.0
info:
  title: WASATEXT
  description: API for WASATEXT messaging platform
  version: 1.0.0
...
security:
  - WASATextAuth: []
paths:
...
  /users/{userId}/conversations/{conversationId}/messages/{messageId}:
    delete:
      summary: Delete message
      description: From the list of sent message, the user can delete one of it.
      operationId: deleteMessage
      tags:
        - "Message"
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/conversationId"
        - $ref: "#/components/parameters/MessageId"
      responses:
        "204":
          $ref: "#/components/responses/DeleteComplete"
        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /users/{userId}/groups:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      summary: View all groups
      description: Returns the list of groups
      operationId: viewGroups
      tags:
        - "Group"
      parameters:
        - in : query
          name : nameGroup
          schema:
            type: string
            minLength: 0
            maxLength: 255
            pattern: "^[a-zA-Z0-9_ -]+$"
            example: capodanno2026
          description: Search group by name. If this field is empty, returns all groups.
          required: false
      responses:
        "200":
          description: OK. Return The list
          content:
            application/json:
              schema:
                description: List of groups
                properties:
                  groups:
                    description: Groups schema
                    type: array
                    minItems: 0
                    maxItems: 2048
                    items:
                      $ref: "#/components/schemas/Group"
                example:
                  groups:
                    - groupId: 288344
                      name: "capodanno2026"
                      upload:
                        photoId: 88332
                        url: "https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg"
                        mime: "image/jpg"
                        width: 1024
                        height: 720
                      participants:
                        - userId: 10012
                          username: "Simone"
                        - userId: 77772
                          username: "Alessio"
    post:
      summary: Create new group
      description: Create a new group conversation.
      operationId: createGroup
      tags:
        - "Group"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Request schema
              properties:
                name:
                  description: Group name
                  type: string
                  minLength: 0
                  maxLength: 255
                  pattern: "^[a-zA-Z0-9_ -]+$"
                  example: capodanno2026
                participants:
                  description: list of participants of a group
                  type: array
                  minItems: 2
                  maxItems: 256
                  items:
                    $ref: "#/components/schemas/Username"
                  example:
                      - Simone
                      - Luca
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                groupId: 288344
                name: "capodanno2026"

                participants:
                  - userId: 10012
                    username: "Simone"
                  - userId: 77772
                    username: "Alessio"

        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /users/{userId}/groups/{groupId}/members:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/GroupId"
    post:
      summary: Add a user into the group
      description: The user can add another user into the group from their userId.
      operationId: addToGroup
      tags:
        - "Group"

      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Request schema
              properties:
                username:
                  $ref: "#/components/schemas/Username"
              required:
                - username
            example:
              username: "Simone"

        required: true
      responses:
        "201":
          description: User added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                groupId: 288344
                name: "capodanno2026"

                participants:
                  - userId: 10012
                    username: "Simone"
                  - userId: 77772
                    username: "Alessio"

        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      summary: List of members
      description: List all members of group
      operationId: getMembers
      tags:
        - "Group"

      responses:
        "200":
          description: OK. Return list.
          content:
            application/json:
              schema:
                description: List of members belonging to the group
                properties:
                  members:
                    description: array that contains members of a group.
                    type: array
                    minItems: 2
                    maxItems: 255
                    items:
                      $ref : "#/components/schemas/User"
              example:
                members:
                  - userId: 11300
                    username: "Simone"
                  - userId: 11301
                    username: "Luca"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /users/{userId}/groups/{groupId}:
    get:
      summary: Show the conversation
      description: Returns the details of a single conversation identified by conversationId from the user's conversations.
      operationId: getGroup
      tags:
        - "Conversations"
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Returns the chat as a list of messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                groupId: 288344
                name: "capodanno2026"
                upload:
                  url: "https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg"
                  mime: "image/jpg"
                  width: 1024
                  height: 720
                participants:
                  - userId: 10012
                    username: "Simone"
                  - userId: 77772
                    username: "Alessio"


        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /users/{userId}/groups/{groupId}/messages/{messageId}:
    delete:
      summary: Delete message
      description: From the list of sent message, the user can delete one of it.
      operationId: deleteGroupMessage
      tags:
        - "Message"
      parameters:
        - $ref: "#/components/parameters/userId"
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/MessageId"
      responses:
        "204":
          $ref: "#/components/responses/DeleteComplete"
        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  schemas:
...
    UserId:
      description: Operation User's ID
      type: integer
      example: 11300
      readOnly: true
    Username:
      description: User's username
      type: string
      minLength: 1
      maxLength: 255
      pattern: "^[a-zA-Z0-9_ -]+$"
      example: Simone
    User:
      description: User schema
      type: object
      properties:
        userId:
          $ref: "#/components/schemas/UserId"
        username:
          $ref: "#/components/schemas/Username"
      required:
        - userId
        - username
    Group:
      type: object
      description: group schema
      properties:
        groupId:
          $ref: "#/components/schemas/GroupId"
        name:
          description: Group name
          type: string
          minLength: 0
          maxLength: 255
          pattern: "^[a-zA-Z0-9_ -]+$"
          example: capodanno2026
        upload:
          $ref: "#/components/schemas/Upload"
        participants:
          description: list of participants of a group
          type: array
          minItems: 2
          maxItems: 256
          items:
            $ref: "#/components/schemas/User"
          example:
            - userId: 11300
              username: Simone
            - userId: 11301
              username: Luca
    ConversationId:
      description: Unique ID assigned to each conversation resource in the system.
      type: integer
      example: 110056
      readOnly: true
...
    MessageId:
      description: Unique ID assigned to each message resource in the system.
      type: integer
      example: 2635444
      readOnly: true
    GroupId:
      description: Unique ID assigned to each group resource in the system.
      type: integer
      example: 2773354
      readOnly: true
...
    Upload:
      description: upload's schema for responses or view
      type: object
      properties:
        url:
          type: string
          minLength: 1
          maxLength: 255
          pattern: "^[a-zA-Z]+://.+$"
          format: url
          description: Photo's URL to access the image
          example: https://www.stoccolmaaroma.it/wp-content/uploads/2013/07/logo-sapienza.jpg
        mime:
          description: mime represents the type of the Photo(image/png...)
          type: string
          minLength: 0
          maxLength: 255
          pattern: '^[a-zA-Z]+/[a-zA-Z.]+$'
          example: image/jpg
        width:
          description: width of the image
          type: integer
          example: 1024
        height:
          description: height of the image
          type: integer
          example: 720
...
    Error:
      description: Info of an error
      type: object
      properties:
        code:
          description: code of the error
          type: integer
          example: 401
        message:
          description: text of the error
          type: string
          minLength: 0
          maxLength: 2048
          example: Unauthorized Action, request invalid or missing.
      required:
        - code
        - message
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User Identifier
      schema:
        $ref: "#/components/schemas/UserId"
    conversationId:
      name: conversationId
      in: path
      required: true
      description: Chat Identifier
      schema:
        $ref: "#/components/schemas/ConversationId"
    MessageId:
      name: messageId
      in: path
      required: true
      description: Message Identifier
      schema:
        $ref: "#/components/schemas/MessageId"
...
    GroupId:
      name: groupId
      in: path
      required: true
      description: Group Identifier
      schema:
        $ref: "#/components/schemas/GroupId"
  securitySchemes:
    WASATextAuth:
      type: http
      description: Protect operation session token.
      scheme: bearer
      bearerFormat: SessionToken
  responses:
    UnauthorizedError:
      description: Unauthorized Action, request invalid or missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 401
            message: "Unauthorized Action, request invalid or missing."
    InternalError:
      description: An error occurred, please retry.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 500
            message: "An error occurred, please retry."
    NotFoundError:
      description: Resource not found. Check the parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 404
            message: "Resource not found. Check the parameters."
    BadRequest:
      description: Bad Request. Invalid parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 400
            message: "Bad Request. Invalid parameters."
    DeleteComplete:
      description: Delete success.